re-add the header file and fix build
